wip: impl kafka producer

+++ src/main/java/com/jwtly/livemarketdata/core/Price.java
@@ -1,13 +1,12 @@
 package com.jwtly.livemarketdata.core;
 
-import java.math.BigDecimal;
 import java.time.Instant;
 
 public record Price(
         String instrument,
-        BigDecimal bid,
-        BigDecimal ask,
-        BigDecimal volume,
+        double bid,
+        double ask,
+        double volume,
         Instant time
 ) {
 }
